Add models for new letter type
