+++ norent/letter_sending.py
@@ -1,13 +1,18 @@
-from typing import Any, Dict, List
+from typing import List
 from io import BytesIO
 import logging
 from django.http import FileResponse
 from django.utils import timezone
 from django.utils.translation import gettext as _
 from django.db import transaction
-import PyPDF2
 
-from project import slack, locales, common_data
+
+from project import slack, locales
+from project.util.letter_sending import (
+    USPS_TRACKING_URL_PREFIX,
+    render_multilingual_letter,
+    send_pdf_to_landlord_via_lob,
+)
 from project.util.site_util import SITE_CHOICES
 from project.util.demo_deployment import is_not_demo_deployment
 from frontend.static_content import (
@@ -16,8 +21,6 @@ from frontend.static_content import (
     ContentType,
 )
 from users.models import JustfixUser
-from loc.views import render_pdf_bytes
-from project.util import lob_api
 from . import models
 
 
@@ -33,9 +36,6 @@ NORENT_EMAIL_TO_LANDLORD_URL = "letter-email.txt"
 # email to the user.
 NORENT_EMAIL_TO_USER_URL = "letter-email-to-user.html"
 
-# The URL prefix for USPS certified letter tracking.
-USPS_TRACKING_URL_PREFIX = common_data.load_json("loc.json")["USPS_TRACKING_URL_PREFIX"]
-
 logger = logging.getLogger(__name__)
 
 
@@ -46,44 +46,6 @@ def norent_pdf_response(pdf_bytes: bytes) -> FileResponse:
     """
 
     return FileResponse(BytesIO(pdf_bytes), filename="norent-letter.pdf")
-
-
-def send_pdf_to_landlord_via_lob(
-    user: JustfixUser, pdf_bytes: bytes, description: str
-) -> Dict[str, Any]:
-    """
-    Mail the given PDF to the given user's landlord using USPS certified
-    mail, via Lob.  Assumes that the user has a landlord with a mailing
-    address.
-
-    Returns the response from the Lob API.
-    """
-
-    ld = user.landlord_details
-    assert ld.address_lines_for_mailing
-    ll_addr_details = ld.get_or_create_address_details_model()
-    landlord_verification = lob_api.verify_address(**ll_addr_details.as_lob_params())
-    user_verification = lob_api.verify_address(**user.onboarding_info.as_lob_params())
-
-    logger.info(
-        f"Sending {description} to landlord with {landlord_verification['deliverability']} "
-        f"landlord address."
-    )
-
-    return lob_api.mail_certified_letter(
-        description=description,
-        to_address={
-            "name": ld.name,
-            **lob_api.verification_to_inline_address(landlord_verification),
-        },
-        from_address={
-            "name": user.full_legal_name,
-            **lob_api.verification_to_inline_address(user_verification),
-        },
-        file=BytesIO(pdf_bytes),
-        color=False,
-        double_sided=False,
-    )
 
 
 def send_letter_via_lob(letter: models.Letter, pdf_bytes: bytes) -> bool:
@@ -184,23 +146,6 @@ def create_letter(user: JustfixUser, rps: List[models.RentPeriod]) -> models.Let
         letter.rent_periods.set(rps)
 
     return letter
-
-
-def _merge_pdfs(pdfs: List[bytes]) -> bytes:
-    merger = PyPDF2.PdfFileMerger()
-    for pdf_bytes in pdfs:
-        merger.append(PyPDF2.PdfFileReader(BytesIO(pdf_bytes)))
-    outfile = BytesIO()
-    merger.write(outfile)
-    return outfile.getvalue()
-
-
-def render_multilingual_letter(letter: models.Letter) -> bytes:
-    pdf_bytes = render_pdf_bytes(letter.html_content)
-    if letter.localized_html_content:
-        localized_pdf_bytes = render_pdf_bytes(letter.localized_html_content)
-        pdf_bytes = _merge_pdfs([pdf_bytes, localized_pdf_bytes])
-    return pdf_bytes
 
 
 def send_letter(letter: models.Letter):
